Fix crash at CTRL_C

+++ src/cpulimit.c
@@ -88,12 +88,14 @@ int lazy = 0;
 static void quit(int sig)
 {
 	//let all the processes continue if stopped
-	struct list_node *node = NULL;
-	for (node=pgroup.proclist->first; node!= NULL; node=node->next) {
-		struct process *p = (struct process*)(node->data);
-		kill(p->pid, SIGCONT);
-	}
-	close_process_group(&pgroup);
+        if (pgroup.proclist != NULL){
+                struct list_node *node = NULL;
+                for (node=pgroup.proclist->first; node!= NULL; node=node->next) {
+                        struct process *p = (struct process*)(node->data);
+                        kill(p->pid, SIGCONT);
+                }
+                close_process_group(&pgroup);
+        }
 	//fix ^C little problem
 	printf("\r");
 	fflush(stdout);
